データ部分を置くためのプラグインを生成する関数。
